chore: move unlock to widgets
