+++ src/renderer/pages/Governance/ui/Governance.tsx
@@ -18,6 +18,7 @@ import {
 } from '@features/governance';
 import { AddDelegationModal } from '@/widgets/AddDelegationModal/components/AddDelegationModal';
 import { addDelegationModel } from '@/widgets/AddDelegationModal/model/addDelegation';
+import { UnlockModal, unlockAggregate } from '@/widgets/UnlockModal';
 import { governancePageAggregate } from '../aggregates/governancePage';
 
 import { EmptyGovernance } from './EmptyGovernance';
@@ -48,7 +49,7 @@ export const Governance = () => {
             <Plate className="w-[240px] h-[90px] pt-3 px-4 pb-4.5">
               <NetworkSelector />
             </Plate>
-            <Locks />
+            <Locks onClick={unlockAggregate.events.flowStarted} />
             <Delegations onClick={addDelegationModel.events.flowStarted} />
           </div>
 
@@ -85,6 +86,7 @@ export const Governance = () => {
       )}
 
       <AddDelegationModal />
+      <UnlockModal />
     </div>
   );
 };
