+++ src/renderer/widgets/UnlockModal/index.ts
@@ -1,0 +1,2 @@
+export { UnlockModal } from './ui/UnlockModal';
+export { unlockAggregate } from './aggregates/unlock';
