+++ src/renderer/widgets/UnlockModal/ui/UnlockInfo.tsx
@@ -7,9 +7,9 @@ import { Button, Duration, FootnoteText, Icon, Shimmering } from '@shared/ui';
 import { AssetBalance } from '@entities/asset';
 import { AssetFiatBalance } from '@entities/price';
 import { permissionUtils, walletModel } from '@entities/wallet';
-import { unlockAggregate } from '../../aggregates/unlock';
-import { locksModel } from '../../model/locks';
-import { votingAssetModel } from '../../model/votingAsset';
+import { locksModel } from '@features/governance/model/locks';
+import { votingAssetModel } from '@features/governance/model/votingAsset';
+import { unlockAggregate } from '../aggregates/unlock';
 
 export const UnlockInfo = () => {
   const { t } = useI18n();
