+++ src/renderer/widgets/UnlockModal/ui/UnlockModal.tsx
@@ -7,15 +7,15 @@ import { Step, isStep } from '@shared/lib/utils';
 import { BaseModal, HeaderTitleText } from '@shared/ui';
 import { OperationTitle } from '@entities/chain';
 import { OperationResult } from '@entities/transaction';
-import { OperationSign, OperationSubmit } from '../../../operations';
-import { unlockAggregate } from '../../aggregates/unlock';
-import { networkSelectorModel } from '../../model/networkSelector';
+import { networkSelectorModel } from '@features/governance/model/networkSelector';
+import { OperationSign, OperationSubmit } from '@features/operations';
+import { unlockAggregate } from '../aggregates/unlock';
 
 import { Confirmation } from './Confirmation';
 import { UnlockForm } from './UnlockForm';
 import { UnlockInfo } from './UnlockInfo';
 
-export const Unlock = () => {
+export const UnlockModal = () => {
   const { t } = useI18n();
 
   const step = useUnit(unlockAggregate.$step);
@@ -66,7 +66,6 @@ export const Unlock = () => {
       {isStep(step, Step.SELECT) && <UnlockForm onGoBack={() => unlockAggregate.events.stepChanged(Step.INIT)} />}
       {isStep(step, Step.CONFIRM) && <Confirmation onGoBack={() => unlockAggregate.events.stepChanged(Step.SELECT)} />}
       {isStep(step, Step.SIGN) && <OperationSign onGoBack={() => unlockAggregate.events.stepChanged(Step.CONFIRM)} />}
-      {isStep(step, Step.SUBMIT) && <OperationSubmit isOpen={isModalOpen} onClose={closeModal} />}
     </BaseModal>
   );
 };
